+++ CHANGES.md
@@ -13,6 +13,8 @@ Changes in CUPS v2.3b1
 - Dropped RSS subscription management from the web interface (Issue #5012)
 - Bonjour printer sharing now uses the DNS-SD hostname (or ServerName value if
   none is defined) when registering shared printers on the network (Issue #5071)
+- The `ipptool` command now supports writing `ippserver` attributes files
+  (Issue #5093)
 - The `lp` and `lpr` commands now provide better error messages when the default
   printer cannot be found (Issue #5096)
 - The `lpadmin` command now provides a better error message when an unsupported